+++ src/tests/injector.spec.ts
@@ -3,9 +3,10 @@ import {Injector} from "../injector/injector";
 import {Logger} from "../logger/logger";
 import {assert, use} from "chai";
 import * as sinonChai from "sinon-chai";
-import {spy, assert as assertSpy} from "sinon";
+import {assert as assertSpy, spy} from "sinon";
 import {Inject} from "../decorators/inject";
 import {Injectable} from "../decorators/injectable";
+import {Metadata} from "../injector/metadata";
 
 // use chai spies
 use(sinonChai);
@@ -30,7 +31,8 @@ describe("Injector", () => {
       @Inject(Logger)
       private c;
 
-      constructor(private a: Logger, @Inject(Logger) private b) {}
+      constructor(private a: Logger, @Inject(Logger) private b) {
+      }
 
       getA(): Logger {
         return this.a;
@@ -63,7 +65,8 @@ describe("Injector", () => {
       @Inject("logger")
       private c;
 
-      constructor(private a: Logger, @Inject("logger") private b) {}
+      constructor(private a: Logger, @Inject("logger") private b) {
+      }
 
       getA(): Logger {
         return this.a;
@@ -98,5 +101,75 @@ describe("Injector", () => {
     assert.instanceOf(a.getC(), Logger);
   });
 
+
+  it("Make sure that @Injection of abstract implementation always delivers correct Token!", () => {
+
+
+    @Injectable()
+    class ServiceA {
+    }
+
+    @Injectable()
+    class ServiceB {
+    }
+
+    @Injectable()
+    class ServiceC {
+    }
+
+    @Injectable()
+    abstract class AbstractClass {
+
+      @Inject(ServiceA)
+      public serviceAClass: ServiceA;
+
+    }
+
+    @Injectable()
+    class ImplementationA extends AbstractClass {
+
+      @Inject(ServiceB)
+      public config: ServiceB;
+
+      @Inject(ServiceB)
+      public serviceBClass: ServiceB;
+
+    }
+
+    @Injectable()
+    class ImplementationB extends AbstractClass {
+
+      @Inject(ServiceC)
+      public config: ServiceC;
+
+      @Inject(ServiceC)
+      public serviceCClass: ServiceC;
+
+    }
+
+    let providerA = Metadata.verifyProvider(ImplementationA);
+    let providerB = Metadata.verifyProvider(ImplementationB);
+
+    let parent = new Injector();
+    parent.createAndResolve(Metadata.verifyProvider(AbstractClass), [Metadata.verifyProvider(ServiceA)]);
+
+    let injector = new Injector(parent);
+    injector.createAndResolve(providerA, [Metadata.verifyProvider(ServiceB)]);
+    injector.createAndResolve(providerB, [Metadata.verifyProvider(ServiceC)]);
+
+    let instanceA = injector.get(ImplementationA);
+    let instanceB = injector.get(ImplementationB);
+
+    assert.instanceOf(instanceA.serviceAClass, ServiceA);
+    assert.instanceOf(instanceA.config, ServiceB);
+    assert.instanceOf(instanceA.serviceBClass, ServiceB);
+    assert.isUndefined(instanceA.serviceCClass);
+
+    assert.instanceOf(instanceB.serviceAClass, ServiceA);
+    assert.instanceOf(instanceB.config, ServiceC);
+    assert.instanceOf(instanceB.serviceCClass, ServiceC);
+    assert.isUndefined(instanceB.serviceBClass);
+
+  });
 
 });